tdf#108480: Undo from clone formating does not remove vertical alignment
In the related change:
49205be73b16bba2a382169aade1f787bcff54d5
wrong method was used to set cell properties, which does
not handle undo actions, so use the proper function.

Other issue was related to undo handling of vertical text
alignment in general. In some cases undo actions were lost
after user clicked out of the cell. It is because vertical
alignment is not affecting the outliner and so the related
undo actions were just dropped in EndTextEdit() method.

Change-Id: I3a6a95fac711056a658a3dee616733939cd13330
Reviewed-on: https://gerrit.libreoffice.org/38699
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tamás Zolnai <tamas.zolnai@collabora.com>