Resolves: tdf#108292 WalkAndMatchElements: really limit the match
... to the columns queried, not just when entering an mdds node. Otherwise it
would return an unexpected index, plus bailing out early spares unnecessary
comparisons for the rest of a node block.

Regression of

    commit 1f9c0bad0a0048e2468d1c2b2d6467173943b325
    Date:   Fri Aug 16 16:29:38 2013 +0200

that started to use

    commit 51011176a8825a31fba3e8a1ed1135c1141e5763
    Date:   Fri Aug 16 16:29:27 2013 +0200

with its bad implementation.

Just that VLOOKUP on a matrix with a larger block of same typed data as the
query *and* a match in an excess column seems to be rare..

Change-Id: Ia4ef3fd56490de82910d5aa13a84be2de851f9b0