Revert "tdf#108524 sw: attempt to split section frames inside table cells"
This reverts commit c2ed0fa70d67c1d752ed38c0b3ec63db8a202bb0.

It doesn't really work and crashes on ooo61225-1.sxw in

1 in SwFrame::FindTabFrame() (this=0x0) at sw/source/core/inc/frame.hxx:913
2 in SwFrame::GetNextSctLeaf(MakePageType) (this=0x3137130, eMakePage=MAKEPAGE_INSERT) at sw/source/core/layout/sectfrm.cxx:1529
3 in SwFrame::GetLeaf(MakePageType, bool) (this=0x3137130, eMakePage=MAKEPAGE_INSERT, bFwd=true) at sw/source/core/layout/flowfrm.cxx:805
4 in SwFlowFrame::MoveFwd(bool, bool, bool) (this=0x31371d8, bMakePage=true, bPageBreak=false, bMoveAlways=false) at sw/source/core/layout/flowfrm.cxx:1861

The code added in GetNextSctLeaf() looks unfinished to me: it assumes that
something else has added a follow-frame for the SwCellFrame containing
the SwSectionFrame already, but AFAICT the GetNextSctLeaf() function
is responsible for creating that SwCellFrame follow.

The caller (in GetLeaf()) specifically checks for this condition and
avoids calling GetNextCellLeaf().

Change-Id: I51875830771f07f5d2fec293f6063c73fc68d468