tdf#108524 sw: attempt to split section frames inside table cells, take two
Tables-in-sections were already split across multiple pages, but not
secions-in-tables. To be safe still don't allow
sections-in-tables-in-sections, so you can combine these in both orders
now, but not recursively.

To achieve this, relax two "not in table" conditions to just require
"not in a table that is already in a section", and define that in case a
section-in-table is to be split, the follow section frame should be
inserted under the follow of its cell.

With this, finally the section frame in the bugdoc is split into two,
and the second section frame is moved to the second page as expected.

This restores commit c2ed0fa70d67c1d752ed38c0b3ec63db8a202bb0, but this second
take makes sure that we do not crash while laying out ooo61225-1.sxw.
GetNextSctLeaf() used to assume that SwTableFrame::Split() created the
follow cell frames; and when that wasn't the case, it crashed. Still
don't attempt to create the cell frame in GetNextSctLeaf(), but handle
when the cell follow frame is missing, and return early, as it used to
be the case before this commit.

Change-Id: I9dcc76b4c61b39b9d23b140b84420e89cf274cf3
Reviewed-on: https://gerrit.libreoffice.org/39471
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>