i#61225 sw: fix layout loop with growable single-column sections
Commit 624923119c513b6029aca56f26fcc7b864558cee (tdf#108524 sw: allow
move of frame inside section without columns, 2017-06-16) relaxed
lcl_IsInSectionDirectly() used in SwFrame::IsMoveable() to allow move
for all section frame contents if it has a single column.  That looked
safe, as the multiple column case was already allowed.

There is one situation where this still causes a problem: when the
section has a single column and the section frame is growable -- as in
that case we should grow the section frame, not move the contents.

So go back to unconditionally allowing multi-column section contents and
allow single-column section contents only in case the section frame is
now growable.

With this, ooo61225-1.sxw from the crashtesting corpus can be opened
again without a layout loop.

Change-Id: Ib2d3702a33da8e62b9bbf468d558ae16db8aa94b
Reviewed-on: https://gerrit.libreoffice.org/39653
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>