Attempt to blind fix tdf#108612 explicitly checking for clipboard document
It seems that for some yet unknown reason in an (un)certain constellation
ScDocument::IsClipboardSource() returns true because ScModule::GetClipDoc()
does not return the clipboard document but the actual document instead, so
destroying the clipboard document executes ScDocument::ClosingClipboardSource()
and (or in earlier versions) ScDocument::ForgetNoteCaptions() on the actual
document instead.

Check also that ScDocument::bIsClip is set, hopefully curing the symptom, the
real cause is something else.

Change-Id: I5ffd0706592d2737808e7004630a56c53115295b