rhbz#739252 sw: fix crash on split tables inside nested sections
Commit 4f106d58d51089ae597aadf854f78f0d02832220 (tdf#108524 sw: attempt
to split section frames inside table cells, take two, 2017-07-03)
checked for tables in SwFrame::GetNextSctLeaf() when it considered
looking up the next "follow" cell frame.

But this is too general, in practice it is only necessary to look for
follow cell frames in case the frame in question is in a table, but not
in a table-in-section. This at the same time avoids a crash with tables
inside nested sections, as it happens in the bugdoc.

Change-Id: If648cb477be5492c7158f89934435ca7021a6a63
Reviewed-on: https://gerrit.libreoffice.org/39692
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>