tdf#109063 DOCX import: consider wrap space for multi-page floattables
Follow-up to commit aecbd9f4fb5a424880c3cd54d2b2b02141ff5f85 (fdo#68607
bnc#816593 DomainMapperTableHandler: don't always start a frame,
2013-09-03), turns out in case there is little space between the table
and the edge of the body area, then there is no wrapping performed in
Word, so we should not convert to floating table, either.

The limit seems to be 266 twips (mm100 unit is used in the code), and
this seems to be constant: it does not change if both the table and the
page width is changed, nor does it change when the empty paragraph to be
wrapped has a different paragraph mark size.

For the majority of the documents this means no change as usually there
is either no space available for wrapping or there is a lot more
available.

Change-Id: Ibbf7409065ba958854514f23b360be56677c8fe3
Reviewed-on: https://gerrit.libreoffice.org/39808
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>