Resolves: tdf#99071 tree view shows odd text widths when > 100 lines
reverts

commit 959ceb379940bb4da926c6f66527ad84a48715f4
Date:   Thu Dec 12 09:55:35 2013 +0100

    fdo#72125: GetTextWidth() can get very expensive.

    Let's just count an approximate width using a cached value when we have too
    many entries.

The expert dialog got fixed by not populating it with all options on load
and instead by incremental disclosure as the users searches/expands it so
this optimization effort isn't needed

in the meantime there was another problem the above papered over with

commit 1d7e48539b71500d239fb33de711e0a12120dba8
Date:   Sun May 17 22:56:46 2015 +0900

refactor how font, fg. and bg. are applied in widgets/controls

which was fixed (hopefully) in the previous commit

Change-Id: I8383d9cd7a9983a85c7f3acec0281d984c44f9e7
Reviewed-on: https://gerrit.libreoffice.org/39951
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>