tdf#109306: consider percent unit specification for table sizes
According to ECMA-376-1:2016, ST_MeasurementOrPercent (the type of
"w:w" attribute of table sizes) is allowed to have this specification
(and then is expected to be a floating-point value). First edition
of the standard (of 2006) only defined this attribute to contain
int32 value (of fiftieths of percent when holding relative widths).

Unit test included.

Change-Id: I700912e4eb07430e55fe1d169d99e8e7e0e1a00b
Reviewed-on: https://gerrit.libreoffice.org/40361
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>