tdf#89409 Render ruby inside the grid in standard mode.
In standard mode the text grids are rectangle and do not have
dedicated space for ruby text. The ruby text either overlap
with base text of the last line or clipped ( after scrolling ).
Move ruby text inside the grid in standard mode as if there
is no grid, and algin text to the center to keep the original
text-to-text alignment behavior.

Change-Id: Ia75ee016f9d492e4e73da6d7245aaeff40cbe1f3
Reviewed-on: https://gerrit.libreoffice.org/40632
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Mark Hung <marklh9@gmail.com>